Utils and sig name change
