+++ configs/addons/counterstrikesharp/gamedata/gamedata.json
@@ -88,7 +88,7 @@
       "linux": "55 48 8D 15 ? ? ? ? 48 89 E5 41 55 49 89 FD 41 54 49 89 F4"
     }
   },
-  "CCSPlayer_CanAcquire": {
+  "CCSPlayer_ItemServices_CanAcquire": {
     "signatures": {
       "library": "server",
       "windows": "48 8B C4 44 89 40 18 48 89 48 08 55 56",
