--- managed/CounterStrikeSharp.API/Modules/Utils/CanAcquire.cs
@@ -1,22 +1,0 @@
-﻿namespace CounterStrikeSharp.API.Modules.Utils;
-
-public enum AcquireMethod : int
-{
-    PickUp = 0,
-    Buy,
-};
-
-public enum AcquireResult : int
-{
-    Allowed = 0,
-    InvalidItem,
-    AlreadyOwned,
-    AlreadyPurchased,
-    ReachedGrenadeTypeLimit,
-    ReachedGrenadeTotalLimit,
-    NotAllowedByTeam,
-    NotAllowedByMap,
-    NotAllowedByMode,
-    NotAllowedForPurchase,
-    NotAllowedByProhibition,
-};
